% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_addresses.R
\name{get_addresses}
\alias{get_addresses}
\title{Get the list of addresses of the specified territory}
\usage{
get_addresses(id)
}
\arguments{
\item{id}{a territory identifier from \strong{ATO_Id} in \strong{get_territories}}
}
\description{
Returns a data frame with information about territories in the
specified region.
}
\details{
The data frame contains the following variables

\itemize{
\item \strong{ATO_Id} - a territory identifier
\item \strong{Geon_Id} - a street id
\item \strong{Geon_Name} - a name of the street
\item \strong{Geon_OldNames} - old names of the street
\item \strong{Bld_ID} - a building identifier
\item \strong{Bld_Area} - a building area
\item \strong{Bld_PS} - a polling station which the building belongs to
\item \strong{Bld_Flats} - a number of flats in the building
\item \strong{Bld_Ind} - an address index
\item \strong{Bld_Korp} - an address block
\item \strong{Bld_Num} - an address number
}
}
