% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_districts.R
\name{get_districts}
\alias{get_districts}
\title{Get the list of election districts}
\usage{
get_districts()
}
\description{
Returns a data frame with IDs, names, shorts name and central localities
of Ukrainian regions.
}
\details{
The data frame contains the following variables

\itemize{
\item \strong{Region_Id} - a region identifier
\item \strong{Area_Num} - a district number
\item \strong{Area_Center} - a central locality of the district
\item \strong{Area_Desc} - borders of the district
\item \strong{Area_OvkAdr} - an address of the district election
election commission
\item \strong{Area_OvkLocation} - a place where the district election
election commission is located
}
}
