% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drv_entities.R
\name{get_drv_entites}
\alias{get_drv_entites}
\title{Get the list of organizations of Ukrainian State Register of Voters}
\usage{
get_drv_entites()
}
\description{
Returns a data frame with IDs of regions, IDs, names, numbers, addresses
and contacts of entities of Ukrainian State Register of Voters
}
\details{
The data frame contains the following variables

\itemize{
\item \strong{Region_Id} - a region identifier
\item \strong{Organ_Id} - an entity identifier
\item \strong{Organ_Num} - an entity number
\item \strong{Organ_Name} - an entity name
\item \strong{Organ_Adr} - an entity address
\item \strong{Organ_Tel} - entity contacts
}
}
