% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polling_station.R
\name{get_polling_stations}
\alias{get_polling_stations}
\title{Get the list of polling stations of the specified district}
\usage{
get_polling_stations(ps)
}
\arguments{
\item{ps}{a district number from \strong{Area_Num} in \strong{get_districts}}
}
\description{
Returns a data frame with information about polling stations in the
specified district.
}
\details{
The data frame contains the following variables

\itemize{
\item \strong{Region_Id} - a region identifier
\item \strong{PS_Area} - a district number
\item \strong{PS_Num} - a polling station number
\item \strong{PS_Type} - a polling station type
\item \strong{PS_Desc} - a polling station borders
\item \strong{PS_CommissionAdr} - a polling station address
\item \strong{PS_PlaceVotingAdr} - a polling station location
\item \strong{PS_PlaceVotingLocation} - a polling station voting address
\item \strong{PS_GeoData} - a polling station polygon (GeoJSON)
\item \strong{PS_GeoDVK} - coordinates of the polling station address
\item \strong{PS_GeoPG} - coordinates of the polling station voting address
\item \strong{PS_Size} - a polling station size
}
}
