% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regions.R
\name{get_regions}
\alias{get_regions}
\title{Get the list of regions}
\usage{
get_regions()
}
\description{
Returns a data frame with IDs, names, shorts name and central localities
of Ukrainian regions.
}
\details{
The data frame contains the following variables

\itemize{
\item \strong{Region_Id} - a region identifier
\item \strong{Region_Num} - a region number
\item \strong{Region_Name} - a region name
\item \strong{Region_Short} - a region shortened name
\item \strong{Region_Center} - a central locality of the region
}
}
